LilyGo T-CAN485 sample update (#39)
While I'm a developer it still took me a bit of time to get going with
the LilyGo T-CAN485. Hope it will takes less time of others with these
hints.

So making the LilyGo T-CAN485 sample easier for first timers:

- The reamde.md recommends to disable WiFi powersave, also added that in
the sample.
- The readme doesn't talk about a static IP, while it's quite practical
to have that.
- I was a bit in the unclear that target address needs to go to home
assistant. Didn't expect the UDP listener in HA. Hinted now that the
target IP goes to HA when HA is used.
- Quite some people will use this with Home Assistant where the default
UDP port is 9999. Also changed to that port in the sample. Also see
issue #35.
- Added some comments around the read and write ports and their
defaults. Makes it a bit easier to do the Home Assistant Nibe setup.